Server: Better handling of network/usb disconnections
Server now stops when communication fails, and leaves destruction
up to the monitor which gets device disconnects from the USB manager.

This way, we clearly distinct between USB and network failure.

The server will just stop in case of any failure, leaving it up
to outside to try and restart the server